+++ src/polytomy.h
@@ -12,7 +12,7 @@ private:
     std::vector<std::string> terminals_to_prune_;
     
 public:
-    Polytomy (const int& seed);
+    Polytomy (const long int& seed);
     void sample_polytomies (Tree * tr);
 };
